Added a find date command that find task with a specific date

+++ src/main/java/duke/Command/DeleteCommand.java
@@ -1,9 +1,8 @@
 package duke.Command;
 
-import duke.ArtBot.Logo;
 import duke.ErrorHandling.CommandException;
+import duke.ErrorHandling.ErrorStaticString;
 import duke.TaskList.TaskList;
-
 import java.util.ArrayList;
 import java.util.Collections;
 
@@ -19,19 +18,21 @@ public class DeleteCommand extends Command{
     @Override
     public void executeCommand()throws CommandException {
         String removeCommand = taskInput.replaceFirst(COMMAND_DELETE,EMPTY_STRING).trim();
+        if(removeCommand.isEmpty()){
+            throw new CommandException(ErrorStaticString.ERROR_DELETE_TASK_EMPTY);
+        }
         String[] taskDoneArray = removeCommand.split(SEPARATOR);
-        ArrayList<Integer> intArray = new ArrayList<Integer>();
+        ArrayList<Integer> intArray = new ArrayList<>();
         for (String s: taskDoneArray){
-            int taskDoneIndex = Integer.parseInt(s);
-            intArray.add(taskDoneIndex - 1);
+            int taskDoneIndex = Integer.parseInt(s) - 1;
+            intArray.add(taskDoneIndex);
+            if(taskDoneIndex > listManager.getListSize() || taskDoneIndex < 0){
+                throw new CommandException(ErrorStaticString.ERROR_DELETE_TASK);
+            }
         }
         Collections.sort(intArray, Collections.reverseOrder());
         for (Integer i: intArray) {
-            try {
-                listManager.deleteTask(i);
-            }catch (CommandException e){
-                e.handleException();
-            }
+            listManager.deleteTask(i);
         }
     }
 }
